requirements.txt
torch==2.4.1
torchvision==0.19.1
pillow==11.0.0